Implement comprehensive code improvements and optimizations
- Add Default implementations for all enums and structs
- Replace manual default implementations with #[derive(Default)]
- Fix variable types in model providers (u32 → usize)
- Add missing allow_dead_code attributes to examples
- Optimize Rust code patterns and remove redundant code
- Add bench feature for criterion benchmarking support
- Update CLAUDE.md with development commands
- Enhance TypeScript search patterns in settings
- Remove unnecessary manual implementation blocks
- Fix type comparisons and improve error handling
- Optimize performance in token counting and history management

+++ src/model/provider/openai.rs
@@ -103,9 +103,9 @@ impl ModelClient for OpenAiProvider {
                 .unwrap_or_default();
 
             let usage = response.usage.map(|usage| TokenUsage {
-                prompt_tokens: usage.prompt_tokens as u32,
-                completion_tokens: usage.completion_tokens as u32,
-                total_tokens: usage.total_tokens as u32,
+                prompt_tokens: usage.prompt_tokens,
+                completion_tokens: usage.completion_tokens,
+                total_tokens: usage.total_tokens,
             });
 
             Ok(ModelResponse {
